Change type to unsigned long long
Change the type of return value of fib_sequence() to unsigned long
long, and change the type of variable sz in client.c to unsigned long
long.

One reason to do that is for representing more number because fibonacci
number is not less than zero.Another reason is for adder which is added
into for even bigger number.